requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pypdf>=6.4.0

# Core AI stack
crewai>=1.6.1
langchain>=1.1.2
langchain-core>=1.1.1
langchain-community>=0.4.1
langchain-google-genai>=3.2.0

# Vector DB + embeddings
chromadb>=1.1.1
sentence-transformers>=3.1.0          # ← THIS WAS MISSING

# Optional but recommended
transformers>=4.45.0
tokenizers>=0.21,<0.23